+++ src/Lakona.Game.Server/Cluster/Membership/Table/MembershipSchemaException.cs
@@ -1,4 +1,14 @@
 namespace Lakona.Game.Cluster.Membership;
 
-public sealed class MembershipSchemaException(string message, Exception innerException)
-    : InvalidOperationException(message, innerException);
+public sealed class MembershipSchemaException : InvalidOperationException
+{
+    public MembershipSchemaException(string message)
+        : base(message)
+    {
+    }
+
+    public MembershipSchemaException(string message, Exception innerException)
+        : base(message, innerException)
+    {
+    }
+}